Move informer event handlers to Scheduler
