+++ pkg/scheduler/factory/factory.go
@@ -20,11 +20,9 @@ package factory
 
 import (
 	"fmt"
-	"reflect"
 	"time"
 
 	v1 "k8s.io/api/core/v1"
-	storagev1 "k8s.io/api/storage/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/fields"
@@ -47,7 +45,6 @@ import (
 	"k8s.io/client-go/tools/record"
 	"k8s.io/klog"
 	podutil "k8s.io/kubernetes/pkg/api/v1/pod"
-	kubeletapis "k8s.io/kubernetes/pkg/kubelet/apis"
 	"k8s.io/kubernetes/pkg/scheduler/algorithm"
 	"k8s.io/kubernetes/pkg/scheduler/algorithm/predicates"
 	"k8s.io/kubernetes/pkg/scheduler/algorithm/priorities"
@@ -66,14 +63,6 @@ import (
 const (
 	initialGetBackoff = 100 * time.Millisecond
 	maximalGetBackoff = time.Minute
-)
-
-var (
-	serviceAffinitySet            = sets.NewString(predicates.CheckServiceAffinityPred)
-	matchInterPodAffinitySet      = sets.NewString(predicates.MatchInterPodAffinityPred)
-	generalPredicatesSets         = sets.NewString(predicates.GeneralPred)
-	noDiskConflictSet             = sets.NewString(predicates.NoDiskConflictPred)
-	maxPDVolumeCountPredicateKeys = []string{predicates.MaxGCEPDVolumeCountPred, predicates.MaxAzureDiskVolumeCountPred, predicates.MaxEBSVolumeCountPred, predicates.MaxCinderVolumeCountPred}
 )
 
 // Binder knows how to write a binding.
@@ -172,8 +161,6 @@ type Configurator interface {
 // configFactory is the default implementation of the scheduler.Configurator interface.
 type configFactory struct {
 	client clientset.Interface
-	// queue for pods that need scheduling
-	podQueue internalqueue.SchedulingQueue
 	// a means to list all known scheduled pods.
 	scheduledPodLister corelisters.PodLister
 	// a means to list all known scheduled pods and pods assumed to have been scheduled.
@@ -226,6 +213,10 @@ type configFactory struct {
 
 	// percentageOfNodesToScore specifies percentage of all nodes to score in each scheduling cycle.
 	percentageOfNodesToScore int32
+
+	bindTimeoutSeconds int64
+	// queue for pods that need scheduling
+	podQueue internalqueue.SchedulingQueue
 }
 
 // ConfigFactoryArgs is a set arguments passed to NewConfigFactory.
@@ -282,104 +273,14 @@ func NewConfigFactory(args *ConfigFactoryArgs) Configurator {
 		hardPodAffinitySymmetricWeight: args.HardPodAffinitySymmetricWeight,
 		disablePreemption:              args.DisablePreemption,
 		percentageOfNodesToScore:       args.PercentageOfNodesToScore,
+		bindTimeoutSeconds:             args.BindTimeoutSeconds,
 	}
-
+	// Setup volume binder
+	c.volumeBinder = volumebinder.NewVolumeBinder(args.Client, args.NodeInformer, args.PvcInformer, args.PvInformer, args.StorageClassInformer, time.Duration(args.BindTimeoutSeconds)*time.Second)
 	c.scheduledPodsHasSynced = args.PodInformer.Informer().HasSynced
-	// scheduled pod cache
-	args.PodInformer.Informer().AddEventHandler(
-		cache.FilteringResourceEventHandler{
-			FilterFunc: func(obj interface{}) bool {
-				switch t := obj.(type) {
-				case *v1.Pod:
-					return assignedPod(t)
-				case cache.DeletedFinalStateUnknown:
-					if pod, ok := t.Obj.(*v1.Pod); ok {
-						return assignedPod(pod)
-					}
-					runtime.HandleError(fmt.Errorf("unable to convert object %T to *v1.Pod for filtering scheduledPod in %T", obj, c))
-					return false
-				default:
-					runtime.HandleError(fmt.Errorf("unable to handle object for filtering scheduledPod in %T: %T", c, obj))
-					return false
-				}
-			},
-			Handler: cache.ResourceEventHandlerFuncs{
-				AddFunc:    c.addPodToCache,
-				UpdateFunc: c.updatePodInCache,
-				DeleteFunc: c.deletePodFromCache,
-			},
-		},
-	)
-	// unscheduled pod queue
-	args.PodInformer.Informer().AddEventHandler(
-		cache.FilteringResourceEventHandler{
-			FilterFunc: func(obj interface{}) bool {
-				switch t := obj.(type) {
-				case *v1.Pod:
-					return !assignedPod(t) && responsibleForPod(t, args.SchedulerName)
-				case cache.DeletedFinalStateUnknown:
-					if pod, ok := t.Obj.(*v1.Pod); ok {
-						return !assignedPod(pod) && responsibleForPod(pod, args.SchedulerName)
-					}
-					runtime.HandleError(fmt.Errorf("unable to convert object %T to *v1.Pod for filtering unscheduledPod in %T", obj, c))
-					return false
-				default:
-					runtime.HandleError(fmt.Errorf("unable to handle object for filtering unscheduledPod in %T: %T", c, obj))
-					return false
-				}
-			},
-			Handler: cache.ResourceEventHandlerFuncs{
-				AddFunc:    c.addPodToSchedulingQueue,
-				UpdateFunc: c.updatePodInSchedulingQueue,
-				DeleteFunc: c.deletePodFromSchedulingQueue,
-			},
-		},
-	)
 	// ScheduledPodLister is something we provide to plug-in functions that
 	// they may need to call.
 	c.scheduledPodLister = assignedPodLister{args.PodInformer.Lister()}
-
-	args.NodeInformer.Informer().AddEventHandler(
-		cache.ResourceEventHandlerFuncs{
-			AddFunc:    c.addNodeToCache,
-			UpdateFunc: c.updateNodeInCache,
-			DeleteFunc: c.deleteNodeFromCache,
-		},
-	)
-
-	args.PvInformer.Informer().AddEventHandler(
-		cache.ResourceEventHandlerFuncs{
-			// MaxPDVolumeCountPredicate: since it relies on the counts of PV.
-			AddFunc:    c.onPvAdd,
-			UpdateFunc: c.onPvUpdate,
-		},
-	)
-
-	// This is for MaxPDVolumeCountPredicate: add/delete PVC will affect counts of PV when it is bound.
-	args.PvcInformer.Informer().AddEventHandler(
-		cache.ResourceEventHandlerFuncs{
-			AddFunc:    c.onPvcAdd,
-			UpdateFunc: c.onPvcUpdate,
-		},
-	)
-
-	// This is for ServiceAffinity: affected by the selector of the service is updated.
-	args.ServiceInformer.Informer().AddEventHandler(
-		cache.ResourceEventHandlerFuncs{
-			AddFunc:    c.onServiceAdd,
-			UpdateFunc: c.onServiceUpdate,
-			DeleteFunc: c.onServiceDelete,
-		},
-	)
-
-	// Setup volume binder
-	c.volumeBinder = volumebinder.NewVolumeBinder(args.Client, args.NodeInformer, args.PvcInformer, args.PvInformer, args.StorageClassInformer, time.Duration(args.BindTimeoutSeconds)*time.Second)
-
-	args.StorageClassInformer.Informer().AddEventHandler(
-		cache.ResourceEventHandlerFuncs{
-			AddFunc: c.onStorageClassAdd,
-		},
-	)
 
 	// Setup cache debugger
 	debugger := cachedebugger.New(
@@ -394,114 +295,7 @@ func NewConfigFactory(args *ConfigFactoryArgs) Configurator {
 		<-c.StopEverything
 		c.podQueue.Close()
 	}()
-
 	return c
-}
-
-// skipPodUpdate checks whether the specified pod update should be ignored.
-// This function will return true if
-//   - The pod has already been assumed, AND
-//   - The pod has only its ResourceVersion, Spec.NodeName and/or Annotations
-//     updated.
-func (c *configFactory) skipPodUpdate(pod *v1.Pod) bool {
-	// Non-assumed pods should never be skipped.
-	isAssumed, err := c.schedulerCache.IsAssumedPod(pod)
-	if err != nil {
-		runtime.HandleError(fmt.Errorf("failed to check whether pod %s/%s is assumed: %v", pod.Namespace, pod.Name, err))
-		return false
-	}
-	if !isAssumed {
-		return false
-	}
-
-	// Gets the assumed pod from the cache.
-	assumedPod, err := c.schedulerCache.GetPod(pod)
-	if err != nil {
-		runtime.HandleError(fmt.Errorf("failed to get assumed pod %s/%s from cache: %v", pod.Namespace, pod.Name, err))
-		return false
-	}
-
-	// Compares the assumed pod in the cache with the pod update. If they are
-	// equal (with certain fields excluded), this pod update will be skipped.
-	f := func(pod *v1.Pod) *v1.Pod {
-		p := pod.DeepCopy()
-		// ResourceVersion must be excluded because each object update will
-		// have a new resource version.
-		p.ResourceVersion = ""
-		// Spec.NodeName must be excluded because the pod assumed in the cache
-		// is expected to have a node assigned while the pod update may nor may
-		// not have this field set.
-		p.Spec.NodeName = ""
-		// Annotations must be excluded for the reasons described in
-		// https://github.com/kubernetes/kubernetes/issues/52914.
-		p.Annotations = nil
-		return p
-	}
-	assumedPodCopy, podCopy := f(assumedPod), f(pod)
-	if !reflect.DeepEqual(assumedPodCopy, podCopy) {
-		return false
-	}
-	klog.V(3).Infof("Skipping pod %s/%s update", pod.Namespace, pod.Name)
-	return true
-}
-
-func (c *configFactory) onPvAdd(obj interface{}) {
-	// Pods created when there are no PVs available will be stuck in
-	// unschedulable queue. But unbound PVs created for static provisioning and
-	// delay binding storage class are skipped in PV controller dynamic
-	// provisioning and binding process, will not trigger events to schedule pod
-	// again. So we need to move pods to active queue on PV add for this
-	// scenario.
-	c.podQueue.MoveAllToActiveQueue()
-}
-
-func (c *configFactory) onPvUpdate(old, new interface{}) {
-	// Scheduler.bindVolumesWorker may fail to update assumed pod volume
-	// bindings due to conflicts if PVs are updated by PV controller or other
-	// parties, then scheduler will add pod back to unschedulable queue. We
-	// need to move pods to active queue on PV update for this scenario.
-	c.podQueue.MoveAllToActiveQueue()
-}
-
-// isZoneRegionLabel check if given key of label is zone or region label.
-func isZoneRegionLabel(k string) bool {
-	return k == kubeletapis.LabelZoneFailureDomain || k == kubeletapis.LabelZoneRegion
-}
-
-func (c *configFactory) onPvcAdd(obj interface{}) {
-	c.podQueue.MoveAllToActiveQueue()
-}
-
-func (c *configFactory) onPvcUpdate(old, new interface{}) {
-	c.podQueue.MoveAllToActiveQueue()
-}
-
-func (c *configFactory) onStorageClassAdd(obj interface{}) {
-	sc, ok := obj.(*storagev1.StorageClass)
-	if !ok {
-		klog.Errorf("cannot convert to *storagev1.StorageClass for storageClassAdd: %v", obj)
-		return
-	}
-
-	// CheckVolumeBindingPred fails if pod has unbound immediate PVCs. If these
-	// PVCs have specified StorageClass name, creating StorageClass objects
-	// with late binding will cause predicates to pass, so we need to move pods
-	// to active queue.
-	if sc.VolumeBindingMode != nil && *sc.VolumeBindingMode == storagev1.VolumeBindingWaitForFirstConsumer {
-		c.podQueue.MoveAllToActiveQueue()
-	}
-}
-
-func (c *configFactory) onServiceAdd(obj interface{}) {
-	c.podQueue.MoveAllToActiveQueue()
-}
-
-func (c *configFactory) onServiceUpdate(oldObj interface{}, newObj interface{}) {
-	c.podQueue.MoveAllToActiveQueue()
-}
-
-func (c *configFactory) onServiceDelete(obj interface{}) {
-	c.podQueue.MoveAllToActiveQueue()
 }
 
 // GetNodeStore provides the cache to the nodes, mostly internal use, but may also be called by mock-tests.
@@ -525,213 +319,6 @@ func (c *configFactory) GetClient() clientset.Interface {
 // GetScheduledPodLister provides a pod lister, mostly internal use, but may also be called by mock-tests.
 func (c *configFactory) GetScheduledPodLister() corelisters.PodLister {
 	return c.scheduledPodLister
-}
-
-func (c *configFactory) addPodToCache(obj interface{}) {
-	pod, ok := obj.(*v1.Pod)
-	if !ok {
-		klog.Errorf("cannot convert to *v1.Pod: %v", obj)
-		return
-	}
-
-	if err := c.schedulerCache.AddPod(pod); err != nil {
-		klog.Errorf("scheduler cache AddPod failed: %v", err)
-	}
-
-	c.podQueue.AssignedPodAdded(pod)
-}
-
-func (c *configFactory) updatePodInCache(oldObj, newObj interface{}) {
-	oldPod, ok := oldObj.(*v1.Pod)
-	if !ok {
-		klog.Errorf("cannot convert oldObj to *v1.Pod: %v", oldObj)
-		return
-	}
-	newPod, ok := newObj.(*v1.Pod)
-	if !ok {
-		klog.Errorf("cannot convert newObj to *v1.Pod: %v", newObj)
-		return
-	}
-
-	if err := c.schedulerCache.UpdatePod(oldPod, newPod); err != nil {
-		klog.Errorf("scheduler cache UpdatePod failed: %v", err)
-	}
-
-	c.podQueue.AssignedPodUpdated(newPod)
-}
-
-func (c *configFactory) addPodToSchedulingQueue(obj interface{}) {
-	if err := c.podQueue.Add(obj.(*v1.Pod)); err != nil {
-		runtime.HandleError(fmt.Errorf("unable to queue %T: %v", obj, err))
-	}
-}
-
-func (c *configFactory) updatePodInSchedulingQueue(oldObj, newObj interface{}) {
-	pod := newObj.(*v1.Pod)
-	if c.skipPodUpdate(pod) {
-		return
-	}
-	if err := c.podQueue.Update(oldObj.(*v1.Pod), pod); err != nil {
-		runtime.HandleError(fmt.Errorf("unable to update %T: %v", newObj, err))
-	}
-}
-
-func (c *configFactory) deletePodFromSchedulingQueue(obj interface{}) {
-	var pod *v1.Pod
-	switch t := obj.(type) {
-	case *v1.Pod:
-		pod = obj.(*v1.Pod)
-	case cache.DeletedFinalStateUnknown:
-		var ok bool
-		pod, ok = t.Obj.(*v1.Pod)
-		if !ok {
-			runtime.HandleError(fmt.Errorf("unable to convert object %T to *v1.Pod in %T", obj, c))
-			return
-		}
-	default:
-		runtime.HandleError(fmt.Errorf("unable to handle object in %T: %T", c, obj))
-		return
-	}
-	if err := c.podQueue.Delete(pod); err != nil {
-		runtime.HandleError(fmt.Errorf("unable to dequeue %T: %v", obj, err))
-	}
-	if c.volumeBinder != nil {
-		// Volume binder only wants to keep unassigned pods
-		c.volumeBinder.DeletePodBindings(pod)
-	}
-}
-
-func (c *configFactory) deletePodFromCache(obj interface{}) {
-	var pod *v1.Pod
-	switch t := obj.(type) {
-	case *v1.Pod:
-		pod = t
-	case cache.DeletedFinalStateUnknown:
-		var ok bool
-		pod, ok = t.Obj.(*v1.Pod)
-		if !ok {
-			klog.Errorf("cannot convert DeletedFinalStateUnknown obj to *v1.Pod: %v", t.Obj)
-			return
-		}
-	default:
-		klog.Errorf("cannot convert to *v1.Pod: %v", t)
-		return
-	}
-
-	if err := c.schedulerCache.RemovePod(pod); err != nil {
-		klog.Errorf("scheduler cache RemovePod failed: %v", err)
-	}
-
-	c.podQueue.MoveAllToActiveQueue()
-}
-
-func (c *configFactory) addNodeToCache(obj interface{}) {
-	node, ok := obj.(*v1.Node)
-	if !ok {
-		klog.Errorf("cannot convert to *v1.Node: %v", obj)
-		return
-	}
-
-	if err := c.schedulerCache.AddNode(node); err != nil {
-		klog.Errorf("scheduler cache AddNode failed: %v", err)
-	}
-
-	c.podQueue.MoveAllToActiveQueue()
-}
-
-func (c *configFactory) updateNodeInCache(oldObj, newObj interface{}) {
-	oldNode, ok := oldObj.(*v1.Node)
-	if !ok {
-		klog.Errorf("cannot convert oldObj to *v1.Node: %v", oldObj)
-		return
-	}
-	newNode, ok := newObj.(*v1.Node)
-	if !ok {
-		klog.Errorf("cannot convert newObj to *v1.Node: %v", newObj)
-		return
-	}
-
-	if err := c.schedulerCache.UpdateNode(oldNode, newNode); err != nil {
-		klog.Errorf("scheduler cache UpdateNode failed: %v", err)
-	}
-
-	// Only activate unschedulable pods if the node became more schedulable.
-	// We skip the node property comparison when there is no unschedulable pods in the queue
-	// to save processing cycles. We still trigger a move to active queue to cover the case
-	// that a pod being processed by the scheduler is determined unschedulable. We want this
-	// pod to be reevaluated when a change in the cluster happens.
-	if c.podQueue.NumUnschedulablePods() == 0 || nodeSchedulingPropertiesChanged(newNode, oldNode) {
-		c.podQueue.MoveAllToActiveQueue()
-	}
-}
-
-func nodeSchedulingPropertiesChanged(newNode *v1.Node, oldNode *v1.Node) bool {
-	if nodeSpecUnschedulableChanged(newNode, oldNode) {
-		return true
-	}
-	if nodeAllocatableChanged(newNode, oldNode) {
-		return true
-	}
-	if nodeLabelsChanged(newNode, oldNode) {
-		return true
-	}
-	if nodeTaintsChanged(newNode, oldNode) {
-		return true
-	}
-	if nodeConditionsChanged(newNode, oldNode) {
-		return true
-	}
-
-	return false
-}
-
-func nodeAllocatableChanged(newNode *v1.Node, oldNode *v1.Node) bool {
-	return !reflect.DeepEqual(oldNode.Status.Allocatable, newNode.Status.Allocatable)
-}
-
-func nodeLabelsChanged(newNode *v1.Node, oldNode *v1.Node) bool {
-	return !reflect.DeepEqual(oldNode.GetLabels(), newNode.GetLabels())
-}
-
-func nodeTaintsChanged(newNode *v1.Node, oldNode *v1.Node) bool {
-	return !reflect.DeepEqual(newNode.Spec.Taints, oldNode.Spec.Taints)
-}
-
-func nodeConditionsChanged(newNode *v1.Node, oldNode *v1.Node) bool {
-	strip := func(conditions []v1.NodeCondition) map[v1.NodeConditionType]v1.ConditionStatus {
-		conditionStatuses := make(map[v1.NodeConditionType]v1.ConditionStatus, len(conditions))
-		for i := range conditions {
-			conditionStatuses[conditions[i].Type] = conditions[i].Status
-		}
-		return conditionStatuses
-	}
-	return !reflect.DeepEqual(strip(oldNode.Status.Conditions), strip(newNode.Status.Conditions))
-}
-
-func nodeSpecUnschedulableChanged(newNode *v1.Node, oldNode *v1.Node) bool {
-	return newNode.Spec.Unschedulable != oldNode.Spec.Unschedulable && newNode.Spec.Unschedulable == false
-}
-
-func (c *configFactory) deleteNodeFromCache(obj interface{}) {
-	var node *v1.Node
-	switch t := obj.(type) {
-	case *v1.Node:
-		node = t
-	case cache.DeletedFinalStateUnknown:
-		var ok bool
-		node, ok = t.Obj.(*v1.Node)
-		if !ok {
-			klog.Errorf("cannot convert DeletedFinalStateUnknown obj to *v1.Node: %v", t.Obj)
-			return
-		}
-	default:
-		klog.Errorf("cannot convert to *v1.Node: %v", t)
-		return
-	}
-
-	if err := c.schedulerCache.RemoveNode(node); err != nil {
-		klog.Errorf("scheduler cache RemoveNode failed: %v", err)
-	}
 }
 
 // Create creates a scheduler with the default algorithm provider.
@@ -966,16 +553,6 @@ func (c *configFactory) getPluginArgs() (*PluginFactoryArgs, error) {
 		VolumeBinder:                   c.volumeBinder,
 		HardPodAffinitySymmetricWeight: c.hardPodAffinitySymmetricWeight,
 	}, nil
-}
-
-// assignedPod selects pods that are assigned (scheduled and running).
-func assignedPod(pod *v1.Pod) bool {
-	return len(pod.Spec.NodeName) != 0
-}
-
-// responsibleForPod returns true if the pod has asked to be scheduled by the given scheduler.
-func responsibleForPod(pod *v1.Pod, schedulerName string) bool {
-	return schedulerName == pod.Spec.SchedulerName
 }
 
 // assignedPodLister filters the pods returned from a PodLister to
